LPS-93662 Move to where it's used.
